Inject panel with ddos scenario

+++ PunchAction/PunchInject.py
@@ -3,6 +3,8 @@ import threading
 from Utils import Constants
 from Utils import FileHelper
 from Utils import SSHHelper
+from Utils import StringHelper
+from Utils import Commands
 
 
 class PunchInject:
@@ -13,21 +15,21 @@ class PunchInject:
         self.client.connectToPunchPlatform()
 
     def injectNoise(self):
-        self.client.sendCommand(
-            Constants.env_variables + Constants.noise_path)
-        #Save PID for each injection process to kill them if wished
-        pids = str.split(self.client.result, '\n')
-        print(pids)
-        FileHelper.FileHelper().saveToPIDFile(pids[1:len(pids)])
+        t = threading.Thread(target=self.client.sendCommand, args=(Constants.env_variables + Constants.noise_path,))
+        t.start()
         return
 
     def stopNoise(self):
-        #Get PID from .pid and kill them
-        pids = FileHelper.FileHelper().getSavedPids()
+        #Get PID by sending ps command ans kill them
+        pids = Commands.Commands().getInjectorPids()
         for pid in pids:
+            if pid == '':
+                continue
             t = threading.Thread(target=self.client.sendCommand, args=(Constants.env_variables + "kill -9 " + pid,))
             t.start()
         return
 
     def injectDDOS(self):
+        t = threading.Thread(target=self.client.sendCommand, args=(Constants.env_variables + Constants.ddos_path,))
+        t.start()
         return
